feat: logout message & demo work on
