+++ budgetplanner/home/urls.py
@@ -1,7 +1,8 @@
 # Django Imports
 from django.urls import path
-from .views import home
+from .views import home, demo_dashboard_view
 
 urlpatterns = [
     path('', home, name='home'),
+    path('demo/', demo_dashboard_view, name='demo_dashboard'),
 ]
